+++ lib/libXft/src/xftextent.c
@@ -100,12 +100,12 @@ XftGlyphExtents (Display	    *dpy,
 		y += xftg->metrics.yOff;
 	    }
 	}
-	extents->x = -overall_left;
-	extents->y = -overall_top;
-	extents->width = overall_right - overall_left;
-	extents->height = overall_bottom - overall_top;
-	extents->xOff = x;
-	extents->yOff = y;
+	extents->x = (short)(-overall_left);
+	extents->y = (short)(-overall_top);
+	extents->width = (unsigned short)(overall_right - overall_left);
+	extents->height = (unsigned short)(overall_bottom - overall_top);
+	extents->xOff = (short)x;
+	extents->yOff = (short)y;
     }
     if (glyphs_loaded)
 	_XftFontManageMemory (dpy, pub);
@@ -127,7 +127,7 @@ XftTextExtents8 (Display	    *dpy,
 	glyphs = glyphs_local;
     else
     {
-	glyphs = malloc (len * sizeof (FT_UInt));
+	glyphs = malloc ((size_t)len * sizeof (FT_UInt));
 	if (!glyphs)
 	{
 	    memset (extents, '\0', sizeof (XGlyphInfo));
@@ -155,7 +155,7 @@ XftTextExtents16 (Display	    *dpy,
 	glyphs = glyphs_local;
     else
     {
-	glyphs = malloc (len * sizeof (FT_UInt));
+	glyphs = malloc ((size_t)len * sizeof (FT_UInt));
 	if (!glyphs)
 	{
 	    memset (extents, '\0', sizeof (XGlyphInfo));
@@ -183,7 +183,7 @@ XftTextExtents32 (Display	    *dpy,
 	glyphs = glyphs_local;
     else
     {
-	glyphs = malloc (len * sizeof (FT_UInt));
+	glyphs = malloc ((size_t)len * sizeof (FT_UInt));
 	if (!glyphs)
 	{
 	    memset (extents, '\0', sizeof (XGlyphInfo));
@@ -217,7 +217,7 @@ XftTextExtentsUtf8 (Display	    *dpy,
     {
 	if (i == size)
 	{
-	    glyphs_new = malloc (size * 2 * sizeof (FT_UInt));
+	    glyphs_new = malloc ((size_t)size * 2 * sizeof (FT_UInt));
 	    if (!glyphs_new)
 	    {
 		if (glyphs != glyphs_local)
@@ -225,7 +225,7 @@ XftTextExtentsUtf8 (Display	    *dpy,
 		memset (extents, '\0', sizeof (XGlyphInfo));
 		return;
 	    }
-	    memcpy (glyphs_new, glyphs, size * sizeof (FT_UInt));
+	    memcpy (glyphs_new, glyphs, (size_t)size * sizeof (FT_UInt));
 	    size *= 2;
 	    if (glyphs != glyphs_local)
 		free (glyphs);
@@ -261,7 +261,7 @@ XftTextExtentsUtf16 (Display		*dpy,
     {
 	if (i == size)
 	{
-	    glyphs_new = malloc (size * 2 * sizeof (FT_UInt));
+	    glyphs_new = malloc ((size_t)size * 2 * sizeof (FT_UInt));
 	    if (!glyphs_new)
 	    {
 		if (glyphs != glyphs_local)
@@ -269,7 +269,7 @@ XftTextExtentsUtf16 (Display		*dpy,
 		memset (extents, '\0', sizeof (XGlyphInfo));
 		return;
 	    }
-	    memcpy (glyphs_new, glyphs, size * sizeof (FT_UInt));
+	    memcpy (glyphs_new, glyphs, (size_t)size * sizeof (FT_UInt));
 	    size *= 2;
 	    if (glyphs != glyphs_local)
 		free (glyphs);
